requirements.txt
fastapi
uvicorn
redis
python-dotenv
langchain
pydantic
tqdm
transformers
